+++ tests/travis_test.sh
@@ -15,3 +15,4 @@ coverage run -m compile_definitions
 coverage run --append -m nose2 -vv
 COOLAMQP_FORCE_SELECT_LISTENER=1 coverage run --append -m nose2 -vv
 coverage run --append -m stress_tests
+COOLAMQP_FORCE_SELECT_LISTENER=1 coverage run --append -m stress_tests